+++ packages/ragbits-core/CHANGELOG.md
@@ -1,8 +1,8 @@
 # CHANGELOG
 
 ## Unreleased
-- Add support for tokenization of given token in LLM class for openai and hf models
 
+- Add token id lookup for LLMs (#513)
 - Allow custom source and element classes to be auto-imported (#343)
 - Make HuggingFaceSource row optional (#529)
 - Allow Source to be constructed from config (#529)